Fix an issue caused by underscores and hyphens at the end of names
Because the code checks for any number of underscores and hyphens
at the end of a name, it will prevent a user with name foobar from
joining if a player with the name foobar_ exists. This allows any
player to prevent another player from logging in by joining with
the same name as the other player, with an underscore added. The
same issue arises if a player exists with the underscore added at
the beginning rather than the end.

This commit changes the regular expression to allow foobar to log
in even when _foobar or foobar_ exists.